refactor(database_utils): move stuff to new crate
