+++ Cargo.toml
@@ -32,6 +32,7 @@ convert_case = "=0.6.0"
 dotenvy_macro = "=0.15.7"
 enum_meta = "=0.7.0"
 itertools = "=0.13.0"
+markdown = "=1.0.0-alpha.18"
 nanoid = "=0.4.0"
 rust_decimal = "=1.35.0"
 rust_decimal_macros = "=1.35.0"
