+++ apps/backend/src/exporter.rs
@@ -24,12 +24,12 @@ use struson::writer::{JsonStreamWriter, JsonWriter};
 use tokio::fs::File;
 use tokio_util::codec::{BytesCodec, FramedRead};
 use traits::AuthProvider;
-use utils::{entity_in_collections, user_measurements_list, IsFeatureEnabled, TEMP_DIR};
-
-use crate::{
-    app_utils::{get_review_export_item, review_by_id, workout_details},
-    background::ApplicationJob,
+use utils::{
+    entity_in_collections, get_review_export_item, review_by_id, user_measurements_list,
+    workout_details, IsFeatureEnabled, TEMP_DIR,
 };
+
+use crate::background::ApplicationJob;
 
 #[derive(Debug, Serialize, Deserialize, SimpleObject, Clone)]
 struct ExportJob {
