+++ crates/utils/database/src/lib.rs
@@ -1,12 +1,34 @@
-use async_graphql::Result;
+use std::sync::Arc;
+
+use application_utils::GraphqlRepresentation;
+use async_graphql::{Error, Result};
+use common_models::IdAndNamedObject;
 use database_models::{
     collection, collection_to_entity,
-    prelude::{Collection, CollectionToEntity, UserMeasurement, UserToCollection},
-    user_measurement, user_to_collection,
+    prelude::{
+        Collection, CollectionToEntity, Review, User, UserMeasurement, UserToCollection, Workout,
+    },
+    user, user_measurement, user_to_collection, workout,
 };
+use dependent_models::UserWorkoutDetails;
+use file_storage_service::FileStorageService;
 use fitness_models::UserMeasurementsListInput;
 use itertools::Itertools;
-use sea_orm::{ColumnTrait, DatabaseConnection, EntityTrait, QueryFilter, QueryOrder, QueryTrait};
+use markdown::to_html as markdown_to_html;
+use media_models::{ImportOrExportItemRating, ImportOrExportItemReview, ReviewItem};
+use rust_decimal_macros::dec;
+use sea_orm::{
+    ColumnTrait, DatabaseConnection, EntityTrait, ModelTrait, QueryFilter, QueryOrder, QueryTrait,
+};
+use user_models::UserReviewScale;
+
+pub async fn user_by_id(db: &DatabaseConnection, user_id: &String) -> Result<user::Model> {
+    User::find_by_id(user_id)
+        .one(db)
+        .await
+        .unwrap()
+        .ok_or_else(|| Error::new("No user found"))
+}
 
 pub async fn user_measurements_list(
     db: &DatabaseConnection,
@@ -63,4 +85,106 @@ pub async fn entity_in_collections(
         .unwrap();
     let resp = mtc.into_iter().flat_map(|(_, b)| b).collect_vec();
     Ok(resp)
+}
+
+pub fn get_review_export_item(rev: ReviewItem) -> ImportOrExportItemRating {
+    let (show_season_number, show_episode_number) = match rev.show_extra_information {
+        Some(d) => (Some(d.season), Some(d.episode)),
+        None => (None, None),
+    };
+    let podcast_episode_number = rev.podcast_extra_information.map(|d| d.episode);
+    let anime_episode_number = rev.anime_extra_information.and_then(|d| d.episode);
+    let manga_chapter_number = rev.manga_extra_information.and_then(|d| d.chapter);
+    ImportOrExportItemRating {
+        review: Some(ImportOrExportItemReview {
+            visibility: Some(rev.visibility),
+            date: Some(rev.posted_on),
+            spoiler: Some(rev.is_spoiler),
+            text: rev.text_original,
+        }),
+        rating: rev.rating,
+        show_season_number,
+        show_episode_number,
+        podcast_episode_number,
+        anime_episode_number,
+        manga_chapter_number,
+        comments: match rev.comments.is_empty() {
+            true => None,
+            false => Some(rev.comments),
+        },
+    }
+}
+
+pub async fn review_by_id(
+    db: &DatabaseConnection,
+    review_id: String,
+    user_id: &String,
+    respect_preferences: bool,
+) -> Result<ReviewItem> {
+    let review = Review::find_by_id(review_id).one(db).await?;
+    match review {
+        Some(r) => {
+            let user = r.find_related(User).one(db).await.unwrap().unwrap();
+            let rating = match respect_preferences {
+                true => {
+                    let preferences = user_by_id(db, user_id).await?.preferences;
+                    r.rating.map(|s| {
+                        s.checked_div(match preferences.general.review_scale {
+                            UserReviewScale::OutOfFive => dec!(20),
+                            UserReviewScale::OutOfHundred => dec!(1),
+                        })
+                        .unwrap()
+                        .round_dp(1)
+                    })
+                }
+                false => r.rating,
+            };
+            Ok(ReviewItem {
+                id: r.id,
+                posted_on: r.posted_on,
+                rating,
+                is_spoiler: r.is_spoiler,
+                text_original: r.text.clone(),
+                text_rendered: r.text.map(|t| markdown_to_html(&t)),
+                visibility: r.visibility,
+                show_extra_information: r.show_extra_information,
+                podcast_extra_information: r.podcast_extra_information,
+                anime_extra_information: r.anime_extra_information,
+                manga_extra_information: r.manga_extra_information,
+                posted_by: IdAndNamedObject {
+                    id: user.id,
+                    name: user.name,
+                },
+                comments: r.comments,
+            })
+        }
+        None => Err(Error::new("Unable to find review".to_owned())),
+    }
+}
+
+pub async fn workout_details(
+    db: &DatabaseConnection,
+    file_storage_service: &Arc<FileStorageService>,
+    user_id: &String,
+    workout_id: String,
+) -> Result<UserWorkoutDetails> {
+    let maybe_workout = Workout::find_by_id(workout_id.clone())
+        .filter(workout::Column::UserId.eq(user_id))
+        .one(db)
+        .await?;
+    match maybe_workout {
+        None => Err(Error::new(
+            "Workout with the given ID could not be found for this user.",
+        )),
+        Some(e) => {
+            let collections =
+                entity_in_collections(db, user_id, None, None, None, None, Some(workout_id))
+                    .await?;
+            let details = e.graphql_representation(file_storage_service).await?;
+            Ok(UserWorkoutDetails {
+                details,
+                collections,
+            })
+        }
+    }
 }
